김다영 / 채팅창 작업중 / 20250226

+++ src/main/java/com/beauty1nside/chat/mapper/ChatMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ChatMapper {
   List<EmpDTO> empList(Long companyNum);
   RoomDTO getChatRoom(RoomDTO roomDTO);
-  int insertChatRoom(RoomDTO roomDTO);
+  Long insertChatRoom(RoomDTO roomDTO);
   List<MessageDTO> getMsgList(Long roomId);
   
 }
